requirements.txt
nextcord==-U
python-dotenv